% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revgeo.R
\name{revgeo}
\alias{revgeo}
\title{Reverse Geocoding with R, Google Maps, and Photon
This function enables reverse geocoding with Google Maps and the Photon API
You must provide a valid Google Maps API to reverse geocode with the Google Maps API
Google Maps is faster and more accurate than Photon, but there's a 2,500 item a day limit}
\usage{
revgeo(longitude, latitude, google = NULL, API = NULL, output = NULL,
  item = NULL)
}
\arguments{
\item{longitude}{Valid longitude coordinate;  e.g., -86.46444}

\item{latitude}{Valid latitude coordunate; e.g., 33.94954}

\item{google}{Defaults to NULL, which specifies Photon API. Enter TRUE to select Google Maps API.}

\item{API}{Defaults to NULL. Enter a valid Google Maps API key to use Google Maps API.}

\item{output}{Defaults to NULL, which returns a reverse geocoded address as a string.  Valid options: 'hash', which returns a hashed string, and 'frame', which returns a dataframe.}

\item{item}{Defaults to NULL, which returns all results.  When output is set to 'hash' or 'frame', this argument let's you specify which portion of the address you want returned; e.g., 'zip' for zipcode. Options include 'housenumber', 'street', 'city', 'county', 'state', and 'country'.}
}
\description{
Reverse Geocoding with R, Google Maps, and Photon
This function enables reverse geocoding with Google Maps and the Photon API
You must provide a valid Google Maps API to reverse geocode with the Google Maps API
Google Maps is faster and more accurate than Photon, but there's a 2,500 item a day limit
}
\examples{
revgeo(longitude=longitude, latitude=latitude, google=NULL, API=NULL, output=NULL, item=NULL)
revgeo(longitude=-86.46222, latitude=33.94954, google=TRUE, API='your API', output='hash', item='zip')
}
\author{
Michael Hudecheck, \email{michael.hudecheck@gess.ethz.ch}
}
\keyword{geocode}
\keyword{reverse}
