% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revgeo.R
\name{revgeo}
\alias{revgeo}
\title{Reverse Geocoding with R, Google Maps, and Photon.}
\usage{
revgeo(longitude, latitude, google = NULL, API = NULL, output = NULL,
  item = NULL)
}
\arguments{
\item{longitude}{Valid longitude coordinate;  e.g., -86.46444}

\item{latitude}{Valid latitude coordunate; e.g., 33.94954}

\item{google}{Defaults to NULL, which specifies Photon API. Enter TRUE to select Google Maps API.}

\item{API}{Defaults to NULL. Enter a valid Google Maps API key to use Google Maps API.}

\item{output}{Defaults to NULL, which returns a reverse geocoded address as a string.  Valid options: 'hash', which returns a hashed string, and 'frame', which returns a dataframe.}

\item{item}{Defaults to NULL, which returns all results.  When output is set to 'hash' or 'frame', this argument let's you specify which portion of the address you want returned; e.g., 'zip' for zipcode. Options include 'housenumber', 'street', 'city', 'county', 'state', and 'country'.}
}
\description{
This function allows you to use the Google Maps API and the Photon API to reverse geocode coordinate pairs.
}
\examples{
revgeo(longitude=-86.46222, latitude=33.94954, google=TRUE, API='your API', output='hash', item='zip')
}
\author{
Michael Hudecheck, \email{michael.hudecheck@gess.ethz.ch}
}
\keyword{geocode}
\keyword{reverse}
